tests: move to c++
