+++ tests/functional/cuda_check.cc
