+++ tests/functional/nccl_connection.cc
@@ -8,7 +8,7 @@
 
 #include "config.h"
 
-#include "test-common.h"
+#include "test-common.hpp"
 
 int main(int argc, char* argv[])
 {
