+++ tests/functional/test-common.hpp
