+++ tests/unit/Makefile.am
@@ -7,7 +7,7 @@
 AM_CPPFLAGS = -I$(top_srcdir)/include
 LDADD = $(top_builddir)/src/libinternal_net_plugin.la
 
-noinst_HEADERS = test-common.h
+noinst_HEADERS = test-common.hpp
 
 noinst_PROGRAMS = \
 	deque \
@@ -21,12 +21,13 @@ noinst_PROGRAMS = \
 
 TESTS = $(noinst_PROGRAMS)
 
-show_tuner_decisions_SOURCES = show_tuner_decisions.c
+show_tuner_decisions_SOURCES = show_tuner_decisions.cc
 show_tuner_decisions_LDADD = $(top_builddir)/src/libinternal_tuner_plugin.la
 
-deque_SOURCES = deque.c
-freelist_SOURCES = freelist.c
-msgbuff_SOURCES = msgbuff.c
-scheduler_SOURCES = scheduler.c
-ep_addr_list_SOURCES = ep_addr_list.c
-mr_SOURCES = mr.c
+idpool_SOURCES = idpool.cc
+deque_SOURCES = deque.cc
+freelist_SOURCES = freelist.cc
+msgbuff_SOURCES = msgbuff.cc
+scheduler_SOURCES = scheduler.cc
+ep_addr_list_SOURCES = ep_addr_list.cc
+mr_SOURCES = mr.cc
